+++ src/core/client/systems/vehicle.ts
@@ -11,8 +11,11 @@ import { getClosestVectorByPos } from '../../shared/utility/vector';
 import { KeybindController } from '../events/keyup';
 import { isAnyMenuOpen } from '../utility/menus';
 import { Timer } from '../utility/timers';
-import { WheelMenu } from '../utility/wheelMenu';
+import { WheelMenu, IWheelItem } from '../utility/wheelMenu';
 import { playAnimation } from './animations';
+
+import './push';
+import { PushVehicle } from './push';
 
 let interval;
 
@@ -98,82 +101,6 @@ export class VehicleController {
         native.setPedConfigFlag(alt.Player.local.scriptID, PedConfigFlag.DisableStoppingVehEngine, true);
     }
 
-    static pushVehicle(vehicle: alt.Vehicle) {
-        const [_null, min, max] = native.getModelDimensions(vehicle.model);
-        native.disableControlAction(0, 32, true);
-        native.disableControlAction(0, 34, true);
-        native.disableControlAction(0, 9, true);
-
-        if (!native.isEntityAttachedToEntity(alt.Player.local.scriptID, vehicle.scriptID)) {
-            native.attachEntityToEntity(
-                alt.Player.local.scriptID,
-                vehicle.scriptID,
-                native.getPedBoneIndex(alt.Player.local.scriptID, 6286),
-                0,
-                min.y - 0.3,
-                min.z + 1,
-                0,
-                0,
-                0,
-                false,
-                false,
-                true,
-                false,
-                0,
-                true
-            );
-        }
-
-        const isPlayingAnim = native.isEntityPlayingAnim(
-            alt.Player.local.scriptID,
-            'missfinale_c2ig_11',
-            'pushcar_offcliff_m',
-            3
-        );
-
-        // S
-        if (native.isControlJustReleased(0, 8)) {
-            if (interval) {
-                Timer.clearInterval(interval);
-                interval = null;
-            }
-
-            native.detachEntity(vehicle.scriptID, false, false);
-            native.detachEntity(alt.Player.local.scriptID, false, false);
-            alt.emitServer(VEHICLE_EVENTS.STOP_PUSH);
-            return;
-        }
-
-        // W
-        if (native.isDisabledControlPressed(0, 32)) {
-            if (!isPlayingAnim) {
-                playAnimation(
-                    'missfinale_c2ig_11',
-                    'pushcar_offcliff_m',
-                    AnimationFlags.REPEAT | AnimationFlags.ENABLE_PLAYER_CONTROL
-                );
-            }
-
-            native.setVehicleForwardSpeed(vehicle.scriptID, 1);
-
-            // A
-            if (native.isDisabledControlPressed(0, 34)) {
-                native.taskVehicleTempAction(alt.Player.local.scriptID, vehicle.scriptID, 11, 1000);
-            }
-
-            // D
-            if (native.isDisabledControlPressed(0, 9)) {
-                native.taskVehicleTempAction(alt.Player.local.scriptID, vehicle.scriptID, 10, 1000);
-            }
-
-            if (native.hasEntityCollidedWithAnything(vehicle.scriptID)) {
-                native.setVehicleOnGroundProperly(vehicle.scriptID, 0);
-            }
-        } else {
-            native.clearPedTasks(alt.Player.local.scriptID);
-        }
-    }
-
     static emitOptions() {
         if (isAnyMenuOpen()) {
             return;
@@ -184,42 +111,21 @@ export class VehicleController {
             return;
         }
 
-        const options = [];
-        const isPushing = alt.Player.local.hasSyncedMeta(PLAYER_SYNCED_META.PUSHING_VEHICLE);
+        const options: Array<IWheelItem> = [];
 
-        if (!isPushing) {
-            options.push({
-                name: 'Push',
-                callback: () => {
-                    if (interval) {
-                        Timer.clearInterval(interval);
-                        interval = null;
-                    }
-
-                    alt.emitServer(VEHICLE_EVENTS.PUSH);
-                    interval = Timer.createInterval(
-                        () => {
-                            VehicleController.pushVehicle(vehicle);
-                        },
-                        0,
-                        'vehicle.ts - pushing'
-                    );
-                }
-            });
-        } else {
-            options.push({
-                name: 'Stop Push',
-                callback: () => {
-                    if (interval) {
-                        alt.clearInterval(interval);
-                        interval = null;
-                    }
-
-                    native.detachEntity(vehicle.scriptID, false, false);
-                    native.detachEntity(alt.Player.local.scriptID, false, false);
-                    alt.emitServer(VEHICLE_EVENTS.STOP_PUSH);
-                }
-            });
+        if (!alt.Player.local.vehicle) {
+            if (!PushVehicle.isPushing() && native.getVehicleDoorLockStatus(vehicle.scriptID) !== 2) {
+                options.push({
+                    name: 'Push',
+                    callback: PushVehicle.start,
+                    data: [vehicle]
+                });
+            } else {
+                options.push({
+                    name: 'Stop Push',
+                    callback: PushVehicle.serverStop
+                });
+            }
         }
 
         WheelMenu.create('Vehicle Options', options);