+++ src/core/shared/enums/playerSynced.ts
@@ -1,6 +1,5 @@
 export enum PLAYER_SYNCED_META {
     WANTED_LEVEL = 'player-wanted',
     POSITION = 'player-position',
-    PING = 'player-ping',
-    PUSHING_VEHICLE = 'player-pushing-vehicle'
+    PING = 'player-ping'
 }